feat(wallet-api): atomic money integrity + server-to-server bank orchestration
- authoritative locked BALANCE + @Version on WALLET_USER (Flyway V4 + ledger backfill); balance read from column, only SUCCESS folds
- BigDecimal money end-to-end (DECIMAL(15,2)); positive-amount/scale + sufficient-funds guards on every flow
- one @Transactional(rollbackFor=Exception) posting via WalletLedgerDAO under PESSIMISTIC_WRITE; typed ApiException control flow
- fix self-transfer (userId .equals), deterministic cashback (2% cap 100) + points (floor/100), locked redeem
- identity derived from JWT on all money endpoints (closes IDOR); amounts moved to request body
- BankClient server-to-server (verify/credit/card-debit/netbanking-debit) with compensation + idempotency keys
- MoneyTxRunner: bounded retry on lock contention (fresh tx per attempt) -> 409 if exhausted
- GlobalExceptionHandler maps ConcurrencyFailure->409 and malformed body->400 (no more 500s)

+++ amigowalletbackend/src/main/java/com/amigowallet/dto/UserProfileResponse.java
@@ -1,5 +1,6 @@
 package com.amigowallet.dto;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 
@@ -17,7 +18,7 @@ public record UserProfileResponse(
 		String emailId,
 		String mobileNumber,
 		UserStatus userStatus,
-		Double balance,
+		BigDecimal balance,
 		Integer rewardPoints,
 		List<CardSummary> cards) {
